Server/host_server/llm/llm_client.py
from dotenv import load_dotenv
import os
from openai import OpenAI

load_dotenv()

api_key = os.getenv("OPENAI_API_KEY")

if not api_key:
    raise ValueError("OPENAI_API_KEY not found in host_server/.env")

client = OpenAI(api_key=api_key)