Server/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
email-validator
openai
fastmcp
requests
certifi
python-jose[cryptography]
passlib[bcrypt]
fastapi-users